requirements.txt
flask>=2.3
groq>=0.9.0
python-dotenv>=1.0
